finish version for first iteration

+++ src/Actions/CancelAppointment.php
@@ -7,8 +7,12 @@ use mindtwo\Appointable\Models\Appointment;
 
 class CancelAppointment
 {
-    public function __invoke(Appointment $appointment): void
+    public function __invoke(Appointment $appointment, bool $force = false): void
     {
+        if (! $force && $appointment->status !== null) {
+            return;
+        }
+
         $appointment->delete();
 
         // dispatch event
